# BACK/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CardResponse": {
				"properties": {
					"lista": {
						"type": "string"
					},
					"conteudo": {
						"type": "string"
					},
					"titulo": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"lista",
					"conteudo",
					"titulo",
					"id"
				],
				"type": "object"
			},
			"Field": {
				"properties": {
					"field": {
						"type": "string"
					},
					"constraints": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"required": [
					"field",
					"constraints"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidationErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"fields": {
						"items": {
							"$ref": "#/components/schemas/Field"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"fields"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CardRequest": {
				"properties": {
					"titulo": {
						"type": "string"
					},
					"conteudo": {
						"type": "string"
					},
					"lista": {
						"type": "string"
					}
				},
				"required": [
					"titulo",
					"conteudo",
					"lista"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			},
			"LoginResponse": {
				"properties": {
					"expiresIn": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"expiresIn",
					"token"
				],
				"type": "object"
			},
			"LoginRequest": {
				"properties": {
					"login": {
						"type": "string"
					},
					"senha": {
						"type": "string"
					}
				},
				"required": [
					"login",
					"senha"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "oauth2",
				"flows": {}
			}
		}
	},
	"info": {
		"title": "BACK",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/cards": {
			"get": {
				"operationId": "GetCards",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CardResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateNewCard",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CardResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CardRequest"
							}
						}
					}
				}
			}
		},
		"/cards/{cardId}": {
			"put": {
				"operationId": "UpdateCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CardResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CardRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CardResponse"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}